Hyundai: add missing FW for 2022 Elantra SEL (#25472)
* added missing FW for 2022 Elantra SEL

I connected the Comma 3 with Hyundai K Harness and the vehicle was unknown. I worked my way through fingerprint 2.0 and compared the FW I had with this file and added the missing firmware to the Elantra_2021 section. I then went for a drive and it worked as intended. A number of the fwVersion were already there and so I've only added the ones that were missing.

* removed erroneous entries

removed entries that were not labeled Hyundai that I had added.

Former-commit-id: c870d0cbdeacc3ab6c3e460fdc484c1961a5b2a0